lib: use size_t for printk return and bool instead of uint8_t

+++ src/lib/inc/printk.h
@@ -18,6 +18,6 @@
 
 #include <bao.h>
 
-unsigned long printk(const char *fmt, ...);
+size_t printk(const char *fmt, ...);
 
 #endif /* __PRINTK_H */
